+++ internal/types/studies.go
@@ -142,41 +142,3 @@ type ContractObjectMetadata struct {
 	// Relationships
 	Contract Contract
 }
-
-func (c Contract) DaysUntilExpiry() *int {
-	if c.ExpiryDate == nil {
-		return nil
-	}
-	expiryDays := int(time.Until(*c.ExpiryDate).Hours() / 24)
-	return &expiryDays
-}
-
-func (c Contract) ShouldNotifyExpiry() bool {
-	daysUntilExpiry := c.DaysUntilExpiry()
-	if daysUntilExpiry == nil {
-		return false
-	} else if *daysUntilExpiry < -90 {
-		return false // don't notify very expired contracts
-	}
-	return *daysUntilExpiry <= 1 || *daysUntilExpiry == 7 || *daysUntilExpiry == 14 || *daysUntilExpiry == 30
-}
-
-// Return the contract with the most urgent expiry notification.
-// Returns nil if there are no contracts that should notify the expiry for
-func (s Study) EarliestExpringContractShouldNotifyExpiry() *Contract {
-	var expiringContract *Contract
-	for _, contract := range s.Contracts {
-		if !contract.ShouldNotifyExpiry() {
-			continue
-		}
-		if expiringContract == nil {
-			expiringContract = &contract
-			continue
-		}
-		daysUntilExpiry := contract.DaysUntilExpiry()
-		if daysUntilExpiry != nil && *daysUntilExpiry < *expiringContract.DaysUntilExpiry() {
-			expiringContract = &contract
-		}
-	}
-	return expiringContract
-}
